chore (refs T22614): Move TocStateMemorizer.js to client/js.
