+++ client/js/bundles/core/publicParagraphDocument.js
@@ -16,7 +16,7 @@ import DpPublicDetailNoMap from '@DpJs/components/statement/DpPublicDetailNoMap'
 import { DpUploadFiles } from '@demos-europe/demosplan-ui'
 import { initialize } from '@DpJs/InitVue'
 import publicStatement from '@DpJs/store/statement/PublicStatement'
-import TocStateMemorizer from '@DemosPlanCoreBundle/lib/TocStateMemorizer'
+import TocStateMemorizer from '@DpJs/lib/core/TocStateMemorizer'
 
 const components = {
   DpPublicDetailNoMap,
